// packages/frontend/src/enums/storage-keys.ts
export enum StorageKeys {
  CLIENT_ID = 'client_id',
  PRIMARY = 'Primary',
}

export enum ValidationMessages {
  SELECT_EVALUATION_TYPE = 'Please select an evaluation type.',
  FILL_SCENARIO_NAMES = 'Please fill all scenario names.',
  UNIQUE_SCENARIO_NAMES = 'Scenario names must be unique.',
  SELECT_APPROACH = 'Please select at least one approach for "{0}" scenario.',
  SELECT_APPROACH1 = 'Please select at least one approach for "Primary" scenario.',
  CLIENT_REQUIRED = 'Client is required.',
}